requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
blobfile==3.0.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.6.0
datatrove==0.5.0
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
evaluate==0.4.3
fastapi==0.116.1
filelock==3.13.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanize==4.12.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.5.1
jsonlines==4.0.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
lm_eval==0.4.8
loguru==0.7.3
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mbstrdecoder==1.1.4
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.5.0
multiprocess==0.70.16
natsort==8.4.0
networkx==3.3
nltk==3.9.1
numexpr==2.11.0
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
objprint==0.3.0
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.10.15
packaging==25.0
pandas==2.3.0
pathvalidate==3.3.1
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.2.0
prettytable==3.16.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyahocorasick==2.1.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.10
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pynvml==12.0.0
pyparsing==3.2.3
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rouge_score==0.1.2
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.47.2
supervision==0.25.1
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
TorchCodec==0.1.0
torchdata==0.11.0
torchinfo==1.8.0
torchmetrics==1.7.3
torchvision==0.21.0+cu124
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.52.4
triton==3.2.0
typepy==1.3.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
viztracer==1.0.3
wandb==0.19.8
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
word2number==1.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
